requirements.txt
boto3
prefect-shell==0.3.1
pytz==2025.2